delete api extra route deleted
