+++ next.config.mjs
@@ -1,4 +1,5 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+};
 
 export default nextConfig;
